Factor out ReplyWriter into its own component
